[feat] 오늘 날짜 컴포넌트 분리
