+++ src/components/Main/MainView.tsx
@@ -5,56 +5,19 @@ import { ReactComponent as Pin } from "../../assests/pin.svg";
 import SideNav from "./SideNav";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
+import Today from "../Common/Today";
 
 const MainView = () => {
-  const today = new Date();
-
-  const year = today.getFullYear();
-  const month = today.getMonth() + 1;
-  const day = today.getDay();
-  const date = today.getDate();
-
   const [sideNavOpen, setSideNavOpen] = useState(false);
   const [isPin, setIsPin] = useState(true);
   const navigate = useNavigate();
-
-  const getKorDay = (day: number) => {
-    switch (day) {
-      case 0:
-        return "일";
-        break;
-      case 1:
-        return "월";
-        break;
-      case 2:
-        return "화";
-        break;
-      case 3:
-        return "수";
-        break;
-      case 4:
-        return "목";
-        break;
-      case 5:
-        return "금";
-        break;
-      case 6:
-        return "토";
-        break;
-
-      default:
-        break;
-    }
-  };
 
   return (
     <MainArea>
       {sideNavOpen && <SideNav setSideNavOpen={setSideNavOpen} />}
       <MainHeader>
         <MenuBtn onClick={() => setSideNavOpen(true)} />
-        <span>
-          {year}. {month}. {date}. ({getKorDay(day)})
-        </span>
+        <Today />
         <Calendar />
       </MainHeader>
       <MapArea></MapArea>
@@ -84,8 +47,6 @@ const MainView = () => {
               </>
             ) : (
               <>
-                <br />
-                <br />
                 아직 오늘의 일기가 없네요!
                 <br />
                 <DiaryBtn onClick={() => navigate("/diary")}>일기쓰기</DiaryBtn>
@@ -137,7 +98,6 @@ const RecordTitle = styled.div`
 `;
 
 const EmptyView = styled.div`
-  position: relative;
   margin-top: 11.5rem;
   display: flex;
   width: 100%;
@@ -152,8 +112,6 @@ const EmptyView = styled.div`
 
 const DiaryBtn = styled.button`
   display: inline-flex;
-  position: fixed;
-  bottom: 1.6rem;
   padding: 0.8rem 5.2rem;
   justify-content: center;
   align-items: center;
